relocate dialpad button to components folder
